// ecommerce/src/modules/auth/dto/auth.dto.d.ts
export declare class RegisterDto {
    email: string;
    password: string;
    fullName: string;
    phone?: string;
}
export declare class LoginDto {
    email: string;
    password: string;
}
export declare class RefreshTokenDto {
    refreshToken: string;
}
export declare class ForgotPasswordDto {
    email: string;
}
export declare class ResetPasswordDto {
    token: string;
    password: string;
}